refactor(ast): extract schema dependency graph into schemaGraph.ts (#3572 1d)
Rename utils/ast.ts to utils/schemaGraph.ts (collectReferencedSchemaNames,
collectUsedSchemaNames, findCircularSchemas, containsCircularRef and their memo
helpers), moving the matching tests alongside it. utils/index.ts re-exports the
schema-graph helpers from their new home. With this, utils/ast.ts is gone and
utils/index.ts is a thin re-export barrel over the six concern-based modules.
