+++ packages/ast/src/utils/index.ts
@@ -3,6 +3,6 @@ export { buildJSDoc, buildList, buildObject, objectKey } from './codegen.ts'
 export { extractStringsFromNodes } from './extractStringsFromNodes.ts'
 export { childName, enumPropName, extractRefName, findDiscriminator, isStringType, resolveGroupType, syncSchemaRef } from './refs.ts'
 export { getNestedAccessor, jsStringEscape, stringify, stringifyObject, toRegExpString, trimQuotes } from './strings.ts'
-export { collectUsedSchemaNames, containsCircularRef, findCircularSchemas } from './ast.ts'
+export { collectUsedSchemaNames, containsCircularRef, findCircularSchemas } from './schemaGraph.ts'
 export { buildGroupParam, buildTypeLiteral, caseParams, resolveParamType } from './operationParams.ts'
 export type { BuildGroupArgs, ParamGroupType } from './operationParams.ts'
